Move session status structs to frontend crate.
This resolves issues with compilation of the dependencies of the main
crate for frontend targets.
