+++ Cargo.toml
@@ -18,6 +18,7 @@ async-trait = "0.1.73"
 base64 = "0.21.4"
 did-web = "0.2.2"
 http = "1.1.0"
+oid4vp-frontend = { path = "oid4vp-frontend" }
 p256 = { version = "0.13.2", features = ["jwk"], optional = true }
 reqwest = { version = "0.12.5", features = ["rustls-tls"], optional = true }
 serde = "1.0.188"
