+++ oid4vp-frontend/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "oid4vp-frontend"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+serde = "1.0"
+serde_json = "1.0"
